\name{epilogi-package}
\alias{epilogi-package}
\docType{package}
\title{
The 'epilogi' Variable Selection Algorithm for Continuous Data.
}
\description{
The '\eqn{\epsilon}pilogi' Variable Selection Algorithm for Continuous Data.
}
\details{
\tabular{lll}{
Package: \tab epilogi\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2024-12-20\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Lakiotaki K., Papadovasilakis Z., Lagani V., Fafalios S., Charonyktakis P., Tsagris M. and Tsamardinos I. (2023). Automated machine learning for Genome Wide Association Studies. Bioinformatics, 39(9): btad545.

Tsagris M., Papadovasilakis Z., Lakiotaki K. and Tsamardinos I. (2022). The \eqn{\gamma}-OMP algorithm for feature selection with application to gene expression data. IEEE/ACM Transactions on Computational Biology and Bioinformatics, 19(2): 1214--1224.
}



